re-factored express app to its own file. index.js starts the server/app
